+++ library/useAES.h
@@ -23,9 +23,9 @@ void aesDecrypt(uint8_t*, uint8_t*, uint8_t*);
     /* Performs a single AES decryption */
 void aesEncrypt(uint8_t*, uint8_t*, uint8_t*);
     /* Performs a single AES Encryption */
-void aesCBCdecrypt(uint8_t*, uint8_t*, uint8_t, uint8_t*, uint8_t*);
+void aesCBCdecrypt(uint8_t*, uint8_t*, uint16_t, uint8_t*, uint8_t*);
     /* Performs complete AES Decryption, CBC Mode */
-void aesCBCencrypt(uint8_t*, uint8_t*, uint8_t*, uint8_t*, uint8_t*);
+void aesCBCencrypt(uint8_t*, uint8_t*, uint16_t*, uint8_t*, uint8_t*);
     /* Performs complete AES Encryption, CBC Mode */
 void simpleDecrypt(uint8_t*, uint8_t*, uint16_t*, uint8_t*);
     /* Performs complete AES Decryption, CBC Mode */
@@ -34,7 +34,7 @@ void simpleEncrypt(uint8_t*, uint8_t*, uint16_t*, uint8_t*);
 
 
 
-void padding(uint8_t*, uint8_t*); 
+void padding(uint8_t*, uint16_t*); 
     /* Pads data until a multipe of 16 */
 
 #endif /* __USEAES_H__ */
